♻️ Refactor: Card components
