+++ client/src/ui/Card/Card.module.scss
@@ -1,5 +1,12 @@
 .card {
-  @apply p-3 relative bg-white bg-opacity-0 rounded-md cursor-pointer;
+  // Size
+  @apply p-3;
+
+  // Position
+  @apply relative;
+
+  // Rest
+  @apply bg-white bg-opacity-0 rounded-md cursor-pointer;
 
   &:hover {
     @apply bg-opacity-5;
@@ -10,36 +17,3 @@
   }
 }
 
-.cardImage {
-  @apply mb-2 relative rounded-md overflow-hidden;
-
-  img {
-    @apply w-full aspect-square;
-  }
-}
-
-.cardImageRounded img {
-  @apply rounded-full;
-}
-
-.name {
-  @apply mb-1 block text-base text-white;
-}
-
-.description {
-  @apply block text-sm text-neutral-400;
-}
-
-.imageIcon {
-  @apply p-3 hidden absolute bottom-2 right-2 size-12 leading-none text-black bg-green-500 rounded-full cursor-pointer;
-
-  transition: all 0.1s ease-out;
-
-  &:hover {
-    @apply bg-green-400 scale-105;
-  }
-
-  &:active {
-    @apply bg-green-600 scale-100;
-  }
-}
