+++ client/src/ui/Card/CardDescription.module.scss
@@ -1,0 +1,18 @@
+.name {
+  // Size
+  @apply mb-1;
+
+  // Display
+  @apply block;
+
+  // Rest
+  @apply text-base text-white;
+}
+
+.description {
+  // Display
+  @apply block;
+
+  // Rest
+  @apply text-sm text-neutral-400;
+}
